chore: remove plugins
