--- bin/cmds/app/plugin.js
@@ -1,9 +1,0 @@
-'use strict';
-
-exports.desc = 'Plugin related commands';
-exports.builder = yargs => {
-	return yargs
-		.commandDir('plugin')
-		.demandCommand()
-		.help()	
-}
