+++ lib/App/index.js
@@ -11,12 +11,7 @@ const Util = require('../..').Util;
 const Log = require('../..').Log;
 const AthomApi = require('../..').AthomApi;
 const Settings = require('../..').Settings;
-const AppPluginCompose = require('../AppPluginCompose');
-const AppPluginZwave = require('../AppPluginZwave');
-const AppPluginZigbee = require('../AppPluginZigbee');
-const AppPluginRF = require('../AppPluginRF');
-const AppPluginLog = require('../AppPluginLog');
-const AppPluginOAuth2 = require('../AppPluginOAuth2');
+const HomeyCompose = require('../HomeyCompose');
 const { AthomAppsAPI } = require('athom-api');
 
 const { getAppLocales } = require('homey-lib');
@@ -50,15 +45,6 @@ const readDirAsync = promisify( fs.readdir );
 const GitCommands = require('../Modules/GitCommands');
 const NpmCommands = require('../Modules/NpmCommands');
 
-const PLUGINS = {
-  'compose': AppPluginCompose,
-  'zwave': AppPluginZwave,
-  'zigbee': AppPluginZigbee,
-  'rf': AppPluginRF,
-  'log': AppPluginLog,
-  'oauth2': AppPluginOAuth2,
-};
-
 const FLOW_TYPES = [ 'triggers', 'conditions', 'actions' ];
 
 class App {
@@ -67,7 +53,6 @@ class App {
 		this.path = path.resolve(appPath);
 		this._app = new HomeyLibApp( this.path );
     this._appJsonPath = path.join( this.path, 'app.json' );
-    this._pluginsPath = path.join( this.path, '.homeyplugins.json');
     this._homeyComposePath = path.join(this.path, '.homeycompose');
 		this._exiting = false;
 		this._std = {};
@@ -622,46 +607,24 @@ class App {
     }
   }
 
-	async preprocess() {
-    	Log(colors.green('✓ Pre-processing app...'));
+  async preprocess() {
+      Log(colors.green('✓ Pre-processing app...'));
 
-		const appJson = await this._getAppJsonFromFolder();
-		if (appJson) {
-			this._isValidAppJson(appJson);
-		} else throw new Error('This app.json is invalid!');
-
-		let plugins = await this._getPlugins();
-		if( plugins.length < 1 ) return;
-
-    Log(colors.green('✓ Running plugins...'));
-
-    for( let i = 0; i < plugins.length; i++ ) {
-      let plugin = plugins[i];
-      let pluginId = plugin.id;
-      let pluginClass = PLUGINS[ pluginId ];
-
-      if( typeof pluginClass !== 'function' )
-        throw new Error(`Invalid plugin: ${pluginId}`);
-
-      Log(colors.green(`✓ Running plugin \`${pluginId}\`...`));
-      let pluginInstance = new pluginClass( this, plugin.options );
-      try {
-        await pluginInstance.run();
-        Log(colors.green(`✓ Plugin \`${pluginId}\` finished`));
-      } catch( err ) {
-        console.trace(err)
-        throw new Error(`Plugin \`${pluginId}\` did not finish:\n${err.message}\n\nAborting.`);
-      }
+    const appJson = await this._getAppJsonFromFolder();
+    if (!appJson) {
+      throw new Error('This app.json is invalid!');
     }
 
+    this._isValidAppJson(appJson);
+
+    HomeyCompose.build({ appPath: this.path });
   }
 
   async version(version) {
-    let hasCompose = await this._hasPlugin('compose');
     let appJsonPath;
     let appJson;
 
-    if( hasCompose ) {
+    if( App.hasHomeyCompose({ appPath: this.path }) ) {
         let appJsonComposePath = path.join(this.path, '.homeycompose', 'app.json');
         let exists = false;
         try {
@@ -1000,66 +963,6 @@ class App {
     Log(colors.white(`\nVisit https://developer.athom.com/apps/app/${appId}/build/${buildId} to publish your app.`));
   }
 
-  async _hasPlugin( pluginId ) {
-    let plugins = await this._getPlugins();
-    for( let i = 0; i < plugins.length; i++ ) {
-      let plugin = plugins[i];
-      if( plugin.id === pluginId ) return true;
-    }
-    return false;
-  }
-
-  async _getPlugins() {
-    try {
-      let plugins = await readFileAsync( this._pluginsPath );
-      return JSON.parse(plugins);
-    } catch( err ) {
-      if( err.code !== 'ENOENT' )
-        throw new Error(`Error in \`.homeyplugins.json\`:\n${err}`);
-    }
-    return [];
-  }
-
-  async addPlugin( pluginId ) {
-    if( !PLUGINS[pluginId] )
-      throw new Error(`Invalid plugin: ${plugin}`);
-    
-    if( await this._hasPlugin(pluginId) ) return;
-    let plugins = await this._getPlugins();
-    plugins.push({
-      id: pluginId
-    });
-    await this._savePlugins( plugins );
-  }
-  
-  async addPluginCLI() {
-    const { pluginId } = await inquirer.prompt([
-      {
-        type: 'list',
-        name: 'pluginId',
-        message: `Select the plugin`,
-        choices: Object.keys(PLUGINS)
-      }
-    ]);
-    
-    await this.addPlugin(pluginId);
-    
-    Log(colors.green(`✓ Added plugin \`${pluginId}\``));
-  }
-  
-  async listPlugins() {
-    const enabledPlugins = await this._getPlugins();
-        
-    Object.keys(PLUGINS).forEach(pluginId => {
-      const enabled = !!enabledPlugins.find(plugin => plugin.id === pluginId);
-      Log(`[${enabled ? 'x' : ' '}] ${pluginId}`)
-    });
-  }
-
-  async _savePlugins( plugins ) {
-    await writeFileAsync( this._pluginsPath, JSON.stringify(plugins, false, 2) );
-  }
-
   _onStd( std ) {
     if( this._exiting ) return;
     if( std.session !== this._session.session ) return;
@@ -1338,7 +1241,7 @@ class App {
   }
 
   async migrateToCompose() {
-    if( await this._hasPlugin('compose') ) throw new Error('You already have the Compose plugin enabled, no need to run this command.');
+    if( App.hasHomeyCompose({ appPath: this.path }) ) throw new Error('You already have the Compose enabled, no need to run this command.');
 
     // Check if the current folder is a git repo. If it is, check for uncommitted changes.
     if ( await GitCommands.isGitInstalled() && await GitCommands.isGitRepo({path: this.path}) ) {
@@ -1499,8 +1402,6 @@ class App {
       );
     } catch( err ) { console.log('Error writing app.json', err) }
 
-    await this.addPlugin('compose');
-
     Log(colors.green(`✓ Successfully migrated app ${appJson.id} to compose`));
 
   }
@@ -1510,7 +1411,7 @@ class App {
       {
         type: 'confirm',
         name: 'switch_compose',
-        message: 'The Homey compose plugin is not detected. Do you want to use Homey compose? It will split the app.json file into separate files for Drivers, Flow Cards and Discovery Strategies.'
+        message: 'The Homey compose is not detected. Do you want to use Homey compose? It will split the app.json file into separate files for Drivers, Flow Cards and Discovery Strategies.'
       }
     )
 
@@ -1595,9 +1496,6 @@ class App {
 
       // TODO pair
 
-			AppPluginZwave.createDriverQuestions(),
-			AppPluginZigbee.createDriverQuestions(),
-			AppPluginRF.createDriverQuestions(),
 			[
 				{
 					type: 'confirm',
@@ -1638,38 +1536,7 @@ class App {
     await copyFileAsync( path.join(templatePath, 'driver.js'), path.join(driverPath, 'driver.js') );
     await copyFileAsync( path.join(templatePath, 'device.js'), path.join(driverPath, 'device.js') );
 
-    if( answers.isZwave ) {
-      await AppPluginZwave.createDriver({
-        driverId,
-        driverPath,
-        answers,
-        driverJson,
-        app: this,
-      });
-    }
-
-    if( answers.isZigbee ) {
-      await AppPluginZigbee.createDriver({
-        driverId,
-        driverPath,
-        answers,
-        driverJson,
-        app: this,
-      });
-    }
-
-    if( answers.isRf ) {
-      await AppPluginRF.createDriver({
-        driverId,
-        driverPath,
-        answers,
-        driverJson,
-        app: this,
-      });
-    }
-
-		let hasCompose = await this._hasPlugin('compose');
-		if( hasCompose ) {
+		if( App.hasHomeyCompose({ appPath: this.path }) ) {
 			if( answers.createDiscovery === true ) {
 				await this.createDiscoveryStrategy();
 			}
@@ -1704,12 +1571,11 @@ class App {
   }
 
   async changeDriverCapabilities() {
-    let hasCompose = await this._hasPlugin('compose');
-    if( !hasCompose ) {
+    if( App.hasHomeyCompose({ appPath: this.path }) === false ) {
       if( await this._askComposeMigration() ) {
         await this.migrateToCompose();
       } else {
-        throw new Error("This command requires the compose plugin to be enabled!");
+        throw new Error("This command requires Homey compose, run `homey app compose` to migrate!");
       }
     }
 
@@ -1825,12 +1691,11 @@ class App {
       this._validateAppJson(appJson);
     }
 
-    let hasCompose = await this._hasPlugin('compose');
-    if( !hasCompose ) {
+    if( App.hasHomeyCompose({ appPath: this.path }) === false ) {
       if( await this._askComposeMigration() ) {
         await this.migrateToCompose();
       } else {
-        throw new Error("This command requires the compose plugin to be enabled!");
+        throw new Error("This command requires Homey compose, run `homey app compose` to migrate!");
       }
     }
 
@@ -2140,12 +2005,11 @@ class App {
       this._validateAppJson(appJson);
     }
 
-    let hasCompose = await this._hasPlugin('compose');
-    if( !hasCompose ) {
+    if( App.hasHomeyCompose({ appPath: this.path }) === false ) {
       if( await this._askComposeMigration() ) {
         await this.migrateToCompose();
       } else {
-        throw new Error("This command requires the compose plugin to be enabled!");
+        throw new Error("This command requires Homey compose, run `homey app compose` to migrate!");
       }
     }
 
@@ -2395,7 +2259,7 @@ class App {
 			{
 				type: 'confirm',
 				name: 'compose',
-				message: 'Use Homey compose plugin?'
+				message: 'Use Homey compose?' // TODO!!!! don't ask, new users don't know what this is
 			},
 			{
 				type: 'confirm',
@@ -2512,13 +2376,7 @@ class App {
       }
     }
 
-    // Now the ap files has been created, we can create a App instance and use that to add plugins.
-    if( answers.compose ) {
-      const app = new App( appPath );
-      app.addPlugin( 'compose' );
-    }
-
-    // Create package lock
+    // Create package.json
     const packageJson = {
       name: answers.id,
       version: answers.version,
@@ -2545,6 +2403,10 @@ class App {
         id: '@types/homey@npm:homey-apps-sdk-v3-types'
       });
     }
+  }
+
+  static hasHomeyCompose({ appPath }) {
+    return fs.existsSync(path.join(appPath, '.homeycompose'));
   }
 
 }
