+++ lib/HomeyCompose/index.js
@@ -1,9 +1,7 @@
 'use strict';
 
 /*
-	Plugin ID: compose
-
-	This plugin generates an app.json file based on scattered files,
+	Homey Compose generates an app.json file based on scattered files,
 	to make it easier to create apps with lots of functionality.
 
 	It finds the following files:
@@ -22,13 +20,6 @@
 	/drivers/<id>/driver.flow.compose.json (object with flow cards, device arg is added automatically)
 	/.homeycompose/locales/en.json
 	/.homeycompose/locales/en.foo.json
-
-	Enable the plugin by adding `{ "id": "compose" }` to your /.homeyplugins.json array
-
-	Plugin options:
-	{
-		"appJsonSpace": 2 | 4 | "\t"
-	}
 */
 
 const fs = require('fs');
@@ -39,24 +30,33 @@ const fse = require('fs-extra');
 const _ = require('underscore');
 const deepmerge = require('deepmerge');
 const objectPath = require('object-path');
+const colors = require('colors');
 const HomeyLib = require('homey-lib');
-
-const AppPlugin = require('../AppPlugin');
 
 const readFileAsync = util.promisify( fs.readFile );
 const writeFileAsync = util.promisify( fs.writeFile );
-const copyFileAsync = util.promisify( fs.copyFile );
+const readdirAsync = util.promisify( fs.readdir );
 
 const deepClone = object => JSON.parse(JSON.stringify(object));
 
 const FLOW_TYPES = [ 'triggers', 'conditions', 'actions' ];
 
-class AppPluginCompose extends AppPlugin {
+class HomeyCompose {
+
+	// Temporary simpler api
+	static async build({ appPath }) {
+		const compose = new HomeyCompose(appPath);
+		await compose.run();
+	}
+
+	constructor(appPath) {
+		this._appPath = appPath;
+	}
+
+	
 
 	async run() {
-
-		this._appPath = this._app.path;
-		this._appPathCompose = path.join(this._app.path, '.homeycompose');
+		this._appPathCompose = path.join(this._appPath, '.homeycompose');
 
 		this._appJsonPath = path.join( this._appPath, 'app.json' );
 		this._appJson = await this._getJsonFile(this._appJsonPath);
@@ -333,7 +333,7 @@ class AppPluginCompose extends AppPlugin {
 			this._appJson.drivers = this._appJson.drivers || [];
 			this._appJson.drivers.push(driverJson);
 
-			this.log(`Added Driver \`${driverId}\``)
+			Log(colors.grey(`Added Driver \`${driverId}\``));
 		}
 	}
 
@@ -358,7 +358,7 @@ class AppPluginCompose extends AppPlugin {
 				this._appJson.signals[ frequency ] = this._appJson.signals[ frequency ] || {};
 				this._appJson.signals[ frequency ][ signalId ] = signal;
 
-				this.log(`Added Signal \`${signalId}\` for frequency \`${frequency}\``)
+				Log(colors.grey(`Added Signal \`${signalId}\` for frequency \`${frequency}\``));
 
 			}
 
@@ -399,7 +399,7 @@ class AppPluginCompose extends AppPlugin {
 		this._appJson.flow[ type ] = this._appJson.flow[ type ] || [];
 		this._appJson.flow[ type ].push( card );
 
-		this.log(`Added FlowCard \`${cardId}\` for type \`${type}\``)
+		Log(colors.grey(`Added FlowCard \`${cardId}\` for type \`${type}\``));
 
 	}
 
@@ -415,7 +415,7 @@ class AppPluginCompose extends AppPlugin {
 			this._appJson.screensavers = this._appJson.screensavers || [];
 			this._appJson.screensavers.push(screensaver);
 
-			this.log(`Added Screensaver \`${screensaver.name}\``)
+			Log(colors.grey(`Added Screensaver \`${screensaver.name}\``));
 		}
 
 	}
@@ -432,7 +432,7 @@ class AppPluginCompose extends AppPlugin {
 			this._appJson.capabilities = this._appJson.capabilities || {};
 			this._appJson.capabilities[ capabilityId ] = capability;
 
-			this.log(`Added Capability \`${capabilityId}\``)
+			Log(colors.grey(`Added Capability \`${capabilityId}\``));
 		}
 
 	}
@@ -448,7 +448,7 @@ class AppPluginCompose extends AppPlugin {
 			this._appJson.discovery = this._appJson.discovery || {};
 			this._appJson.discovery[ strategyId ] = strategy;
 
-			this.log(`Added Discovery Strategy \`${strategyId}\``)
+			Log(colors.grey(`Added Discovery Strategy \`${strategyId}\``));
 		}
 
 	}
@@ -496,8 +496,8 @@ class AppPluginCompose extends AppPlugin {
 		for( let i = 0; i < appLocalesChanged.length; i++ ) {
 			let appLocaleId = appLocalesChanged[i];
 			let appLocale = appLocales[appLocaleId];
-			await writeFileAsync( path.join(appLocalesPath, `${appLocaleId}.json`), JSON.stringify(appLocale, false, this._options.appJsonSpace || 2) )
-			this.log(`Added Locale \`${appLocaleId}\``)
+			await writeFileAsync( path.join(appLocalesPath, `${appLocaleId}.json`), JSON.stringify(appLocale, false, 2) )
+			Log(colors.grey(`Added Locale \`${appLocaleId}\``));
 		}
 
 	}
@@ -519,9 +519,68 @@ class AppPluginCompose extends AppPlugin {
 		let json = JSON.parse(JSON.stringify(this._appJson));
 			json = removeDollarPropertiesRecursive(json);
 
-		await writeFileAsync( this._appJsonPath, JSON.stringify(json, false, this._options.appJsonSpace || 2) );
+		await writeFileAsync( this._appJsonPath, JSON.stringify(json, false, 2) );
+	}
+
+	async _getChildFolders( rootPath ) {
+    let childFolders = [];
+		try {
+			const pathContents = await readdirAsync(rootPath, { withFileTypes: true });
+
+			// Check all paths for dirs
+      for (let i in pathContents) {
+        if (pathContents[i].isDirectory()) childFolders.push(pathContents[i].name);
+      }
+
+      return childFolders;
+		} catch( err ) {
+      return childFolders;
+		}
+	}
+
+	async _getFiles( filesPath ) {
+		try {
+			let files = await readdirAsync( filesPath )
+
+			return files.filter(file => {
+				return file.indexOf('.') !== 0;
+			}).sort((a, b) => {
+				a = path.basename( a, path.extname(a) ).toLowerCase();
+				b = path.basename( b, path.extname(b) ).toLowerCase();
+				return a.localeCompare(b);
+			});
+		} catch( err ) {
+			return [];
+		}
+	}
+
+	async _getJsonFiles( filesPath ) {
+		let result = {};
+		let files = await this._getFiles( filesPath );
+		for( let i = 0; i < files.length; i++ ) {
+			let filePath = files[i];
+			if( path.extname(filePath) !== '.json' ) continue;
+
+			let fileJson = await this._getJsonFile( path.join( filesPath, filePath ) );
+			let fileId = path.basename( filePath, '.json' );
+
+			result[ fileId ] = fileJson;
+
+		}
+		return result;
+	}
+
+	async _getJsonFile( filePath ) {
+		let fileJson = await readFileAsync( filePath );
+		try {
+			fileJson = JSON.parse(fileJson);
+		} catch( err ) {
+			throw new Error(`Error in file ${filePath}\n${err.message}`);
+		}
+
+		return fileJson;
 	}
 
 }
 
-module.exports = AppPluginCompose;
+module.exports = HomeyCompose;
